// types/user.d.ts
import {CognitoUserSession} from 'amazon-cognito-identity-js';
import {Chart} from './';
import {Workout} from './workout';

// USER
export type LoggedInUser = {
    data: CognitoUserSession | null;
    docData: UserDocData | null;
    loggedIn: boolean;
};

export type UserDocData = {
    _id: string;
    username: string;
    country: string;
    dob: string;
    email: string;
    firstName: string;
    profilePhoto: string;
    surname: string;
    gender: string;
    height: number;
    metric: boolean;
    options?: {
        charts?: {
            homepage?: {
                leftRail?: Chart[];
                rightRaight?: Chart[];
            };
            profile?: {
                leftRail?: Chart[];
                rightRaight?: Chart[];
            };
            workout?: {
                leftRail?: Chart[];
                rightRaight?: Chart[];
            };
        };
    };
    weight: number;
    workouts: Workout[];
};

export type UserProfile = {
    _id: string;
    username: string;
    followerCount: number;
    followingCount: number;
    followers: UserReference[];
    options: {
        charts?: {
            profile?: {
                leftRail?: Chart[];
                rightRail?: Chart[];
            };
        };
    };
    isFollowed: boolean;
    isLoggedInUser: boolean;
};

export type UserReference = {
    username: string;
    userId: string;
    createdAt: Date;
    profilePhoto: string;
};

// API
export type GetUserParams = {
    userId: string;
    init: GetUserInit;
};

export type GetUserInit = {
    headers?: {
        Authorization?: string;
    };
    queryStringParameters?: {
        view?: string;
    };
};

export type UpdateUserParams = {
    username: string;
    init: UpdateUserInit;
};

export type UpdateUserInit = {
    headers?: {
        Authorization?: string;
    };
    body: {
        userForm: {
            country?: string;
            dob?: string;
            email?: string;
            firstName?: string;
            profilePhoto?: string;
            surname?: string;
            gender?: string;
            height?: number;
            metric?: boolean;
            options?: {
                charts?: {
                    homepage?: {
                        leftRail?: Chart[];
                        rightRaight?: Chart[];
                    };
                    profile?: {
                        leftRail?: Chart[];
                        rightRaight?: Chart[];
                    };
                    workout?: {
                        leftRail?: Chart[];
                        rightRaight?: Chart[];
                    };
                };
            };
            weight?: number;
        };
    };
};